requirements.txt
flask 
flask-socketio 
azure-cognitiveservices-speech 
requests
eventlet